+++ src/main/kotlin/cl/ravenhill/kernet/functions/activation.kt
@@ -60,11 +60,13 @@ abstract class AbstractActivationFunction<T : TType>(protected val tf: Ops) :
 
 /**
  * Wrapper class for the Sigmoid activation function.
+ *
  * @param T
  *    the type of data used by the operation
  * @property tf
  *    the context for the operations
  * @see [Ops]
+ * @see [Sigmoid]
  */
 class KSigmoid<T : TType>(tf: Ops) : AbstractActivationFunction<T>(tf) {
   override fun setFeatures(x: Operand<T>): KSigmoid<T> {
@@ -80,13 +82,83 @@ class KSigmoid<T : TType>(tf: Ops) : AbstractActivationFunction<T>(tf) {
   }
 }
 
+/**
+ * Wrapper class for the ReLU activation function.
+ *
+ * @param T
+ *    the type of data used by the operation
+ * @property tf
+ *    the context for the operations
+ * @see [Ops]
+ * @see [Relu]
+ */
+class KReLU<T : TType>(tf: Ops) : AbstractActivationFunction<T>(tf) {
+  override fun call(): Relu<T> = tf.nn.relu(features)
+
+  override fun invoke(x: Operand<T>): Relu<T> {
+    setFeatures(x)
+    return call()
+  }
+
+  override fun setFeatures(x: Operand<T>): KReLU<T> {
+    features = x
+    return this
+  }
+}
+
+/**
+ * Wrapper class for the tanh activation function.
+ *
+ * @param T
+ *    the type of data used by the operation
+ * @property tf
+ *    the context for the operations
+ * @see [Ops]
+ */
+class KTanh<T : TType>(tf: Ops) : AbstractActivationFunction<T>(tf) {
+  override fun call(): Tanh<T> = tf.math.tanh(features)
+
+  override fun invoke(x: Operand<T>): Tanh<T> {
+    setFeatures(x)
+    return call()
+  }
+
+  override fun setFeatures(x: Operand<T>): KTanh<T> {
+    features = x
+    return this
+  }
+}
+
+/**
+ * Wrapper class for the Softmax activation function.
+ *
+ * @param T
+ *    the type of data used by the operation
+ * @property tf
+ *    the context for the operations
+ * @see [Ops], [Softmax]
+ */
+class KSoftmax<T : TNumber>(tf: Ops) : AbstractActivationFunction<T>(tf) {
+  override fun call(): Softmax<T> = tf.nn.softmax(features)
+
+  override fun invoke(x: Operand<T>): Softmax<T> {
+    setFeatures(x)
+    return call()
+  }
+
+  override fun setFeatures(x: Operand<T>): KSoftmax<T> {
+    features = x
+    return this
+  }
+}
+
 fun <T : TType> sigmoid(tf: Ops, features: Operand<T>) = KSigmoid<T>(tf).invoke(features)
 
-fun <T : TType> relu(tf: Ops, features: Operand<T>): Relu<T> = tf.nn.relu(features)
+fun <T : TType> relu(tf: Ops, features: Operand<T>) = KReLU<T>(tf).invoke(features)
 
-fun <T : TType> tanh(tf: Ops, features: Operand<T>): Tanh<T> = tf.math.tanh(features)
+fun <T : TType> tanh(tf: Ops, features: Operand<T>) = KTanh<T>(tf).invoke(features)
 
-fun <T : TNumber> softmax(tf: Ops, features: Operand<T>): Softmax<T> = tf.nn.softmax(features)
+fun <T : TNumber> softmax(tf: Ops, features: Operand<T>) = KSoftmax<T>(tf).invoke(features)
 
 fun <T : TType> swish(tf: Ops, features: Operand<T>, beta: Tensor<T>): Operand<T> {
   OperatorContext.setOperatorContext(tf)