+++ CHANGELOG.md
@@ -2,4 +2,6 @@
 
 # Version 0.1
 
+- ``0.0.3.0`` Softmax and tanh functions
+- ``0.0.2.0`` ReLU activation function
 - ``0.0.1.0`` Sigmoid activation function